# artifacts/api-server/src/routes/analytics.ts
import { Router, type IRouter, type Request, type Response } from "express";
import { db } from "@workspace/db";
import {
  leadsTable,
  usersTable,
  communicationsTable,
  emailSendsTable,
  leadStatusHistoryTable,
} from "@workspace/db";
import { eq, and, gte, lte, sql, inArray, isNotNull, desc } from "drizzle-orm";
import { getUserDisplayName, requireUser } from "../lib/authHelpers";
import { activeRepListingCondition, normalizeFundingTimeDays } from "../lib/analyticsHelpers";
import { unassignedInboundLeadCondition } from "../lib/inboundLead";

const router: IRouter = Router();

const FUNNEL_ORDER = [
  "new_lead",
  "contacted",
  "application_received",
  "submitted_to_underwriting",
  "approved",
  "funded",
] as const;

const APPLICATION_STATUSES = ["application_received", "submitted_to_underwriting", "approved", "funded"] as const;
const APPROVAL_STATUSES = ["approved", "funded"] as const;

// Admin-only operational count. Keep this predicate shared with the lead-list
// needsAssignment field so the dashboard cannot drift from the row indicator.
router.get("/analytics/unassigned-inbound-count", async (req: Request, res: Response) => {
  const user = await requireUser(req, res);
  if (!user) return;
  if (user.role !== "admin") {
    return void res.status(403).json({ error: "Forbidden: admins only" });
  }

  const [{ count }] = await db
    .select({ count: sql<number>`cast(count(*) as int)` })
    .from(leadsTable)
    .where(unassignedInboundLeadCondition());

  res.json({ count });
});

function parseDateRange(req: Request): { startDate?: Date; endDate?: Date; repId?: number } {
  const { start_date, end_date, rep_id } = req.query as Record<string, string | undefined>;
  const startDate = start_date ? new Date(start_date) : undefined;
  const endDate = end_date ? new Date(end_date + "T23:59:59.999Z") : undefined;
  const repId = rep_id ? parseInt(rep_id, 10) : undefined;
  return { startDate, endDate, repId };
}

function leadDateWhere(startDate?: Date, endDate?: Date, repId?: number) {
  const clauses = [];
  if (startDate) clauses.push(gte(leadsTable.createdAt, startDate));
  if (endDate) clauses.push(lte(leadsTable.createdAt, endDate));
  if (repId && !isNaN(repId)) clauses.push(eq(leadsTable.assignedRepId, repId));
  return clauses.length ? and(...clauses) : undefined;
}

// GET /analytics/summary
router.get("/analytics/summary", async (req: Request, res: Response) => {
  const user = await requireUser(req, res);
  if (!user) return;

  const { startDate, endDate, repId } = parseDateRange(req);

  // Reps can only see their own data
  const effectiveRepId = user.role === "rep" ? user.id : repId;

  const where = leadDateWhere(startDate, endDate, effectiveRepId);
  const allTimeWhere = user.role === "rep"
    ? eq(leadsTable.assignedRepId, user.id)
    : undefined;

  const [allLeads, leadCountRows, allTimeLeadCountRows, fundedTimeRows, revenueRows] = await Promise.all([
    db
      .select({ status: leadsTable.status, count: sql<number>`cast(count(*) as int)` })
      .from(leadsTable)
      .where(where)
      .groupBy(leadsTable.status),

    // Leads generated is every lead created in the selected range, regardless
    // of source or status. Keep this independent from funnel categorization.
    db
      .select({ count: sql<number>`cast(count(*) as int)` })
      .from(leadsTable)
      .where(where),

    // Used only for the first-run state. Managers should not lose the
    // established-CRM check when changing filters, while reps must not see
    // another user's records in this value.
    db
      .select({ count: sql<number>`cast(count(*) as int)` })
      .from(leadsTable)
      .where(allTimeWhere),

    // Avg time from lead created to funded status in history
    db
      .select({
        avgDays: sql<number>`avg(extract(epoch from (${leadStatusHistoryTable.createdAt} - ${leadsTable.createdAt})) / 86400)`,
      })
      .from(leadStatusHistoryTable)
      .innerJoin(leadsTable, eq(leadStatusHistoryTable.leadId, leadsTable.id))
      .where(
        and(
          eq(leadStatusHistoryTable.toStatus, "funded"),
          sql`${leadStatusHistoryTable.createdAt} > ${leadsTable.createdAt} + interval '5 minutes'`,
          startDate ? gte(leadsTable.createdAt, startDate) : undefined,
          endDate ? lte(leadsTable.createdAt, endDate) : undefined,
          effectiveRepId ? eq(leadsTable.assignedRepId, effectiveRepId) : undefined,
        ),
      ),

    // Total revenue from funded leads in range
    db
      .select({
        totalRevenue: sql<number>`cast(coalesce(sum(${leadsTable.fundedAmount}), 0) as int)`,
      })
      .from(leadsTable)
      .where(and(eq(leadsTable.status, "funded"), where)),
  ]);

  const countByStatus: Record<string, number> = {};
  for (const row of allLeads) {
    if (row.status) countByStatus[row.status] = row.count;
  }

  const totalLeads = leadCountRows[0]?.count ?? 0;
  const allTimeTotalLeads = allTimeLeadCountRows[0]?.count ?? 0;
  const totalApplications = APPLICATION_STATUSES.reduce((a, s) => a + (countByStatus[s] ?? 0), 0);
  const totalApprovals = APPROVAL_STATUSES.reduce((a, s) => a + (countByStatus[s] ?? 0), 0);
  const totalFundings = countByStatus["funded"] ?? 0;
  const totalRevenue = revenueRows[0]?.totalRevenue ?? 0;
  const conversionRate = totalLeads > 0 ? Math.round((totalFundings / totalLeads) * 10000) / 100 : 0;
  const avgFundingTimeDays =
    normalizeFundingTimeDays(fundedTimeRows[0]?.avgDays != null ? Number(fundedTimeRows[0].avgDays) : null);

  res.json({
    totalLeads,
    allTimeTotalLeads,
    totalApplications,
    totalApprovals,
    totalFundings,
    totalRevenue,
    conversionRate,
    avgFundingTimeDays,
  });
});

// GET /analytics/pipeline
router.get("/analytics/pipeline", async (req: Request, res: Response) => {
  const user = await requireUser(req, res);
  if (!user) return;

  const { startDate, endDate, repId } = parseDateRange(req);
  const effectiveRepId = user.role === "rep" ? user.id : repId;
  const where = leadDateWhere(startDate, endDate, effectiveRepId);

  const rows = await db
    .select({ status: leadsTable.status, count: sql<number>`cast(count(*) as int)` })
    .from(leadsTable)
    .where(where)
    .groupBy(leadsTable.status);

  const countByStatus: Record<string, number> = {};
  for (const row of rows) {
    if (row.status) countByStatus[row.status] = row.count;
  }

  const stages = FUNNEL_ORDER.map((status, idx) => {
    const count = countByStatus[status] ?? 0;
    const prevCount = idx > 0 ? (countByStatus[FUNNEL_ORDER[idx - 1]] ?? 0) : null;
    const conversionFromPrevious =
      prevCount !== null && prevCount > 0 ? Math.round((count / prevCount) * 10000) / 100 : null;
    return { status, count, conversionFromPrevious };
  });

  res.json({ stages });
});

type ListAnalyticsRepsDependencies = {
  database?: typeof db;
  authenticate?: typeof requireUser;
};

/**
 * Build the rep analytics handler with production defaults. Keeping the
 * database/auth collaborators injectable lets the API regression exercise the
 * actual HTTP handler and SQL predicate without a second test-only query.
 */
export function createListAnalyticsRepsHandler({
  database = db,
  authenticate = requireUser,
}: ListAnalyticsRepsDependencies = {}) {
  return async (req: Request, res: Response) => {
    const user = await authenticate(req, res);
    if (!user) return;

    if (user.role === "rep") {
      return void res.status(403).json({ error: "Forbidden: managers and admins only" });
    }

    const { startDate, endDate } = parseDateRange(req);

    const leadDateClauses = [];
    if (startDate) leadDateClauses.push(gte(leadsTable.createdAt, startDate));
    if (endDate) leadDateClauses.push(lte(leadsTable.createdAt, endDate));
    const leadWhere = leadDateClauses.length ? and(...leadDateClauses) : undefined;

    const commDateClauses = [];
    if (startDate) commDateClauses.push(gte(communicationsTable.createdAt, startDate));
    if (endDate) commDateClauses.push(lte(communicationsTable.createdAt, endDate));
    const commWhere = commDateClauses.length ? and(...commDateClauses) : undefined;

    const emailDateClauses = [];
    if (startDate) emailDateClauses.push(gte(emailSendsTable.createdAt, startDate));
    if (endDate) emailDateClauses.push(lte(emailSendsTable.createdAt, endDate));
    const emailWhere = emailDateClauses.length ? and(...emailDateClauses) : undefined;

    const [reps, leadCounts, revenueByRep, callCounts, smsCounts, emailCounts] = await Promise.all([
      database
        .select()
        .from(usersTable)
        .where(activeRepListingCondition(database)),

      database
        .select({
          repId: leadsTable.assignedRepId,
          status: leadsTable.status,
          count: sql<number>`cast(count(*) as int)`,
        })
        .from(leadsTable)
        .where(leadWhere)
        .groupBy(leadsTable.assignedRepId, leadsTable.status),

      database
        .select({
          repId: leadsTable.assignedRepId,
          revenue: sql<number>`cast(coalesce(sum(${leadsTable.fundedAmount}), 0) as int)`,
        })
        .from(leadsTable)
        .where(and(eq(leadsTable.status, "funded"), leadWhere))
        .groupBy(leadsTable.assignedRepId),

      database
        .select({
          userId: communicationsTable.userId,
          count: sql<number>`cast(count(*) as int)`,
        })
        .from(communicationsTable)
        .where(
          and(
            eq(communicationsTable.type, "call"),
            eq(communicationsTable.direction, "outbound"),
            commWhere,
          ),
        )
        .groupBy(communicationsTable.userId),

      database
        .select({
          userId: communicationsTable.userId,
          count: sql<number>`cast(count(*) as int)`,
        })
        .from(communicationsTable)
        .where(
          and(
            eq(communicationsTable.type, "sms"),
            eq(communicationsTable.direction, "outbound"),
            commWhere,
          ),
        )
        .groupBy(communicationsTable.userId),

      database
        .select({
          userId: emailSendsTable.userId,
          count: sql<number>`cast(count(*) as int)`,
        })
        .from(emailSendsTable)
        .where(emailWhere)
        .groupBy(emailSendsTable.userId),
    ]);

    const callMap: Record<number, number> = {};
    for (const r of callCounts) if (r.userId) callMap[r.userId] = r.count;

    const smsMap: Record<number, number> = {};
    for (const r of smsCounts) if (r.userId) smsMap[r.userId] = r.count;

    const emailMap: Record<number, number> = {};
    for (const r of emailCounts) if (r.userId) emailMap[r.userId] = r.count;

    const revenueMap: Record<number, number> = {};
    for (const r of revenueByRep) if (r.repId) revenueMap[r.repId] = r.revenue;

    // leadCounts grouped by repId + status
    type LeadCountEntry = { repId: number | null; status: string | null; count: number };
    const leadByRepStatus: Record<number, Record<string, number>> = {};
    for (const r of leadCounts as LeadCountEntry[]) {
      if (!r.repId || !r.status) continue;
      if (!leadByRepStatus[r.repId]) leadByRepStatus[r.repId] = {};
      leadByRepStatus[r.repId][r.status] = r.count;
    }

    const result = reps.map((rep) => {
      const statusMap = leadByRepStatus[rep.id] ?? {};
      const leadsCount = Object.values(statusMap).reduce((a, b) => a + b, 0);
      const applications = APPLICATION_STATUSES.reduce((a, s) => a + (statusMap[s] ?? 0), 0);
      const approvals = APPROVAL_STATUSES.reduce((a, s) => a + (statusMap[s] ?? 0), 0);
      const fundings = statusMap["funded"] ?? 0;
      return {
        repId: rep.id,
        repName: getUserDisplayName(rep),
        leadsCount,
        callsMade: callMap[rep.id] ?? 0,
        smsSent: smsMap[rep.id] ?? 0,
        emailsSent: emailMap[rep.id] ?? 0,
        applications,
        approvals,
        fundings,
        revenue: revenueMap[rep.id] ?? 0,
      };
    });

    res.json(result);
  };
}

// GET /analytics/reps
router.get("/analytics/reps", createListAnalyticsRepsHandler());

// GET /analytics/sources
router.get("/analytics/sources", async (req: Request, res: Response) => {
  const user = await requireUser(req, res);
  if (!user) return;

  const { startDate, endDate, repId } = parseDateRange(req);
  const effectiveRepId = user.role === "rep" ? user.id : repId;

  const clauses = [];
  if (startDate) clauses.push(gte(leadsTable.createdAt, startDate));
  if (endDate) clauses.push(lte(leadsTable.createdAt, endDate));
  if (effectiveRepId) clauses.push(eq(leadsTable.assignedRepId, effectiveRepId));
  const where = clauses.length ? and(...clauses) : undefined;

  const rows = await db
    .select({
      source: leadsTable.leadSource,
      status: leadsTable.status,
      count: sql<number>`cast(count(*) as int)`,
      revenue: sql<number>`cast(coalesce(sum(${leadsTable.fundedAmount}), 0) as int)`,
    })
    .from(leadsTable)
    .where(where)
    .groupBy(leadsTable.leadSource, leadsTable.status);

  const sourceMap: Record<string, { totalCount: number; fundedCount: number; revenue: number }> = {};
  for (const row of rows) {
    const src = row.source ?? "manual";
    if (!sourceMap[src]) sourceMap[src] = { totalCount: 0, fundedCount: 0, revenue: 0 };
    sourceMap[src].totalCount += row.count;
    if (row.status === "funded") {
      sourceMap[src].fundedCount += row.count;
      sourceMap[src].revenue += row.revenue;
    }
  }

  const result = Object.entries(sourceMap).map(([source, { totalCount, fundedCount, revenue }]) => ({
    source,
    leadCount: totalCount,
    fundedCount,
    conversionRate: totalCount > 0 ? Math.round((fundedCount / totalCount) * 10000) / 100 : 0,
    revenue,
  }));

  res.json(result);
});

// GET /analytics/communications
router.get("/analytics/communications", async (req: Request, res: Response) => {
  const user = await requireUser(req, res);
  if (!user) return;

  const { startDate, endDate, repId } = parseDateRange(req);
  const effectiveRepId = user.role === "rep" ? user.id : repId;
  const granularity: "day" | "week" = (req.query["granularity"] as string) === "weekly" ? "week" : "day";

  const clauses = [];
  if (startDate) clauses.push(gte(communicationsTable.createdAt, startDate));
  if (endDate) clauses.push(lte(communicationsTable.createdAt, endDate));
  if (effectiveRepId) clauses.push(eq(communicationsTable.userId, effectiveRepId));
  const where = clauses.length ? and(...clauses) : undefined;

  // granularity is restricted to a fixed whitelist above, so it is safe to inline as a raw
  // SQL literal here. It must NOT be passed as a bound parameter: when the same `sql` fragment
  // is referenced in SELECT, GROUP BY, and ORDER BY, each reference gets its own bind parameter
  // ($1/$2/$3), and Postgres cannot prove those separately-parameterized expressions are
  // identical, causing "column must appear in the GROUP BY clause" errors on every query.
  const truncExpr = sql<string>`date_trunc(${sql.raw(`'${granularity}'`)}, "communications"."created_at")::date::text`;

  const rows = await db
    .select({
      date: truncExpr,
      type: communicationsTable.type,
      count: sql<number>`cast(count(*) as int)`,
    })
    .from(communicationsTable)
    .where(where)
    .groupBy(truncExpr, communicationsTable.type)
    .orderBy(truncExpr);

  // Merge call + sms into same date entry
  const dateMap: Record<string, { date: string; calls: number; sms: number }> = {};
  for (const row of rows) {
    const d = row.date;
    if (!dateMap[d]) dateMap[d] = { date: d, calls: 0, sms: 0 };
    if (row.type === "call") dateMap[d].calls += row.count;
    if (row.type === "sms") dateMap[d].sms += row.count;
  }

  res.json(Object.values(dateMap).sort((a, b) => a.date.localeCompare(b.date)));
});

// GET /analytics/renewals
router.get("/analytics/renewals", async (req: Request, res: Response) => {
  const user = await requireUser(req, res);
  if (!user) return;

  const { repId } = parseDateRange(req);
  const effectiveRepId = user.role === "rep" ? user.id : repId;

  const clauses = [isNotNull(leadsTable.renewalFlaggedAt)];
  if (effectiveRepId && !isNaN(effectiveRepId)) clauses.push(eq(leadsTable.assignedRepId, effectiveRepId));

  const rows = await db.query.leadsTable.findMany({
    where: and(...clauses),
    orderBy: [desc(leadsTable.renewalFlaggedAt)],
    with: { assignedRep: true },
  });

  res.json(
    rows.map((l) => ({
      id: l.id,
      firstName: l.firstName,
      lastName: l.lastName,
      companyName: l.companyName,
      fundedAt: l.fundedAt?.toISOString() ?? null,
      renewalFlaggedAt: l.renewalFlaggedAt?.toISOString() ?? null,
      assignedRepId: l.assignedRepId,
      assignedRepName: l.assignedRep ? getUserDisplayName(l.assignedRep) : null,
      requestedAmount: l.requestedAmount ?? null,
    })),
  );
});

export default router;