# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MBS CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: KPI summary — total leads, applications, approvals, fundings, conversion rate, avg funding time
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: rep_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
        "401":
          description: Unauthorized
  /analytics/unassigned-inbound-count:
    get:
      operationId: getUnassignedInboundCount
      tags: [analytics]
      summary: Count unassigned website and QR-card leads older than 24 hours (admin only)
      responses:
        "200":
          description: Current inbound assignment backlog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnassignedInboundCount"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /analytics/pipeline:
    get:
      operationId: getAnalyticsPipeline
      tags: [analytics]
      summary: Lead count per funnel stage with conversion rates
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: rep_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Pipeline funnel data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsPipeline"
        "401":
          description: Unauthorized
  /analytics/reps:
    get:
      operationId: getAnalyticsReps
      tags: [analytics]
      summary: Per-rep performance metrics (manager/admin only)
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Rep performance table
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepPerformance"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /analytics/sources:
    get:
      operationId: getAnalyticsSources
      tags: [analytics]
      summary: Lead volume and conversion by source
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: rep_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Source analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceAnalytics"
        "401":
          description: Unauthorized
...
  /analytics/communications:
    get:
      operationId: getAnalyticsCommunications
      tags: [analytics]
      summary: Daily or weekly call and SMS volume over the selected period
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: rep_id
          in: query
          schema:
            type: integer
        - name: granularity
          in: query
          schema:
            type: string
            enum: [daily, weekly]
            default: daily
      responses:
        "200":
          description: Communication activity time series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommActivity"
        "401":
          description: Unauthorized
  /analytics/renewals:
    get:
      operationId: getAnalyticsRenewals
      tags: [analytics]
      summary: List funded leads flagged by the renewal radar job as ready to re-fund
      parameters:
        - name: rep_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Renewal-flagged funded leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RenewalOpportunity"
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AnalyticsSummary:
      type: object
      required: [totalLeads, allTimeTotalLeads, totalApplications, totalApprovals, totalFundings, conversionRate]
      properties:
        totalLeads:
          type: integer
          description: Total leads created in the selected date range, regardless of lead source or status
        allTimeTotalLeads:
          type: integer
          description: Unfiltered all-time lead count used only to determine whether the CRM is truly empty
        totalApplications:
          type: integer
        totalApprovals:
          type: integer
        totalFundings:
          type: integer
        totalRevenue:
          type: integer
          description: Sum of fundedAmount across funded leads in the date range
        conversionRate:
          type: number
        avgFundingTimeDays:
          type: ["number", "null"]
    UnassignedInboundCount:
      type: object
      required: [count]
      properties:
        count:
          type: integer
          description: Unassigned website or QR-card leads created more than 24 hours ago
    AnalyticsPipeline:
      type: object
      required: [stages]
      properties:
        stages:
          type: array
          items:
            type: object
            required: [status, count]
            properties:
              status:
                type: string
              count:
                type: integer
              conversionFromPrevious:
                type: ["number", "null"]
    RepPerformance:
      type: object
      required: [repId, repName, leadsCount, callsMade, smsSent, emailsSent, applications, approvals, fundings]
      properties:
        repId:
          type: integer
        repName:
          type: string
        leadsCount:
          type: integer
        callsMade:
          type: integer
        smsSent:
          type: integer
        emailsSent:
          type: integer
        applications:
          type: integer
        approvals:
          type: integer
        fundings:
          type: integer
        revenue:
          type: integer
          description: Sum of fundedAmount for this rep's funded leads in the date range
    SourceAnalytics:
      type: object
      required: [source, leadCount, fundedCount, conversionRate]
      properties:
        source:
          type: string
        leadCount:
          type: integer
        fundedCount:
          type: integer
        conversionRate:
          type: number
        revenue:
          type: integer
          description: Sum of fundedAmount for funded leads from this source in the date range
    CommActivity:
      type: object
      required: [date, calls, sms]
      properties:
        date:
          type: string
        calls:
          type: integer
        sms:
          type: integer
    RenewalOpportunity:
      type: object
      required: [id, fundedAt, renewalFlaggedAt]
      properties:
        id:
          type: integer
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        fundedAt:
          type: ["string", "null"]
          format: date-time
        renewalFlaggedAt:
          type: string
          format: date-time
        assignedRepId:
          type: ["integer", "null"]
        assignedRepName:
          type: ["string", "null"]
        requestedAmount:
          type: ["integer", "null"]
...